+++ src/Products/ZReST/tests/__init__.py
@@ -1,2 +1,0 @@
-""" Unit tests for ZReST product.
-"""